requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
backports.asyncio.runner==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
datasets==4.5.0
dill==0.4.0
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
idna==3.11
iniconfig==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
multiprocess==0.70.18
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
propcache==0.4.1
pyarrow==23.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
tomli==2.4.0
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
xxhash==3.6.0
yarl==1.22.0
